Chapter 3 has been started.
Chp3.QueuesAndCommands - I've learned that work which is the collection of commands are proccessed in CommandBuffer. How to create commandBuffer from commandBufferPool. Also destroying all of them. The point I have been surprised was do not need to free all command buffer when pool is destroyed. Surely, we should stop tracking buffers when its pool is destroyed.