Chapter 3 is finished!
Chp3.QueuesAndCommands - Learned how to recycle used buffers and submit command buffers to work. However, we did not cover it fully yet. For example, how to call one command buffer from another and how to implement fin-grained synchronization between the host and the device and between queues on the device.